wip: move indexing to link time.
